ajuste na nomenclatura + remoção de classes desnecessarias + inclusão do metodo MAIN
