--- main.py
@@ -1,4 +1,0 @@
-import server
-
-if __name__ == "__main__":
-    server.start_server()
