+++ setup.py
@@ -4,9 +4,9 @@ setup(
     name='pomodoro',
     version='0.0.1',
     install_requires=['click', 'requests'],
-    py_modules=['pomodoro', 'interface', 'log', 'win_messages', 'server',],
+    py_modules=['pomodoro', 'pomodoro_core', 'log', 'win_messages', 'server',],
     entry_points='''
         [console_scripts]
-        pomodoro=interface:pomodoro
+        pomodoro=pomodoro:pomodoro
     ''',
 )
